hospital-booking/mcp_server/requirements.txt
# MCP SDK
mcp>=1.0.0

# HTTP Client for FastAPI calls
httpx==0.27.0

# Environment variables
python-dotenv==1.0.0

# Async support
asyncio

# Type hints
typing-extensions==4.9.0
